fixed random del's!

+++ core/framework/ui/games.js
@@ -424,8 +424,9 @@ module.exports = {
                   for (var x = 0; x < all_messages.length; x++)
                     if (!fetched_game_embeds.includes(all_messages[x][0]))
                       try {
-                        if (!all_messages[x][1].author.bot)
-                          all_messages[x][1].delete();
+                        if (local_ui.channel != settings.alert_channel)
+                          if (!all_messages[x][1].author.bot)
+                            all_messages[x][1].delete();
                       } catch {}
                 });
           }, 3000, local_ui, local_interface);
